Add update() in TriMotorDrive.java
update() is a way to let the mode RUN_TO_POSITION move to the right place, and not overshoot.

This is implemented in LeftAuto.java, with the robot moving 10 inches forward.
This is not actually going to be done, because the constant variable  INCHES_PER_ROTATION isn't the right number yet.